+++ defaults.py
@@ -229,6 +229,7 @@ class FlatCAMDefaults:
         "excellon_plot": True,
         "excellon_solid": True,
         "excellon_multicolored": False,
+        "excellon_merge_fuse_tools": True,
         "excellon_format_upper_in": 2,
         "excellon_format_lower_in": 4,
         "excellon_format_upper_mm": 3,